+++ src/app/_home/components/Category/Category/index.ts
@@ -1,0 +1,1 @@
+export { default as Category } from "./Category";
