+++ src/app/_home/components/Category/CategoryList/CategoryList.module.scss
