+++ src/app/_home/components/Category/CategoryList/index.ts
@@ -1,0 +1,1 @@
+export { default as CategoryList } from "./CategoryList";
