+++ src/app/_home/components/Category/index.ts
@@ -1,0 +1,2 @@
+export { Category } from "./Category";
+export { CategoryList } from "./CategoryList";
