+++ src/app/_home/components/Main/index.ts
@@ -1,0 +1,1 @@
+export { default as Main } from "./Main";
