+++ src/app/_home/components/Products/PopularProducts/PopularProducts.tsx
@@ -1,5 +1,5 @@
-import { ProductList } from "@/app/_home/components/Products/ProductList";
-import { list } from "@/app/_home/mock";
+import { ProductList } from "@/_home/components/Products/ProductList";
+import { list } from "@/_home/mock";
 import cn from "@/utils/classNames";
 import styles from "./PopularProducts.module.scss";
