+++ src/app/_home/components/Products/index.ts
@@ -1,0 +1,2 @@
+export { FilteredProducts } from "./FilteredProducts";
+export { PopularProducts } from "./PopularProducts";
