+++ tsconfig.json
@@ -21,7 +21,8 @@
     "paths": {
       "@/*": ["src/*"],
       "@/images/*": ["public/images/*"],
-      "@/fonts/*": ["public/fonts/*"]
+      "@/fonts/*": ["public/fonts/*"],
+      "@/_home/*": ["src/app/_home/*"]
     }
   },
   "include": ["next-env.d.ts", "**/*.ts", "**/*.tsx", ".next/types/**/*.ts"],
